ADD: fix create table for lat/lng in artist;

+++ 01_data_modelling_project/project_template/sql_queries.py
@@ -44,8 +44,8 @@ artist_table_create = ("""
     artist_id text, 
     name text, 
     location text, 
-    latitude int, 
-    longitude int)
+    latitude float8, 
+    longitude float8)
 """)
 
 time_table_create = ("""
